requirements.txt
Flask
slackclient
gevent
